% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageGroups.R
\name{ageGroups}
\alias{ageGroups}
\title{Age groups}
\usage{
ageGroups(name = "esp2013", type = "broad")
}
\arguments{
\item{name}{Dataset - One of "esp2013" or ""wsp2025".}

\item{type}{Choose whether to use broad or narrow age groups.}
}
\value{
List
}
\description{
Age groups
}
