% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsr.R
\name{dsr}
\alias{dsr}
\title{Calculate directly standardised rates}
\usage{
dsr(
  data,
  refdata = standardPopulation("esp2013"),
  event,
  time,
  age = "age_group",
  pop = "pop",
  strata = NULL,
  mp = 1000,
  method = "normal",
  sig = 0.95,
  decimals = 4
)
}
\arguments{
\item{data}{A data frame with the event counts to be standardised.}

\item{refdata}{A data frame representing the standard population. It must contain two columns:
age, with the different age groups (notice that this column name must be the same as
in data, defined by the input age); and pop, with the number of individuals in each corresponding
age group.}

\item{event}{Name of the column in data that corresponds to the event counts.}

\item{time}{Name of the column in data that corresponds to the unit-time (denominator).}

\item{age}{Name of the column in data and refdata that corresponds to age groups.}

\item{pop}{Name of the column in refdata that corresponds to the standard population in each age group.}

\item{strata}{Name of the columns in data for which rates are calculated by.}

\item{mp}{A constant to multiply rates by (e.g. mp = 1000 for rates per 1000).}

\item{method}{Choose between normal, lognormal and gamma confidence intervals
for crude and standardised rates. The default method is normal.}

\item{sig}{The desired level of confidence in computing confidence intervals.
The default is 0.95 for 95 percent CIs.}

\item{decimals}{Round estimates to a desired decimal place.}
}
\description{
Computes crude and directly standardised rates. Rates can be stratified by
variables of interest.
}
\examples{
# An example of calculating directly standardised rates
# Data example is from Table 1 (p.132) of Fundamentals of Epidemiology by Schoenbach, 2000.

# The following table shows the number of deaths, for 5 different age groups,
# in the states of Miami and Alaska:
data <- data.frame(
      state = rep(c('Miami',"Alaska"), c(5,5)),
      age_groups = rep(c('00-14','15-24','25-44','45-64','65+'),2),
      deaths = c(136, 57, 208, 1016, 3605, 59, 18, 37, 90, 81),
      general_population = c(114350,80259,133440,142670,92168,37164,20036,32693,14947,2077))

# We aim to standardise the number of deaths per each state. To do that, we will use the following
# US standard population:
standardised_population <- data.frame(
                            age_groups = c('00-14','15-24','25-44','45-64','65+'),
                            pop = c(23961000,15420000,21353000,19601000,10685000))

# Now we will use the function dsr to calculate the direct standardised rates
# (per 1000 individuals) using a 95\% CI calculated by the gamma method:
my_results <- dsr(data = data,
                  refdata = standardised_population,
                  event = "deaths",
                  time  = "general_population",
                  age   = "age_groups",
                  pop   = "pop",
                  strata = "state",
                  mp = 1000,
                  method = "gamma",
                  sig = 0.95,
                  decimals = 8)
# View results
my_results
}
