% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAgeGroups.R
\name{mergeAgeGroups}
\alias{mergeAgeGroups}
\title{Create new merged age groups}
\usage{
mergeAgeGroups(
  refdata,
  newGroups,
  event = NULL,
  age = "age_group",
  pop = "pop",
  strata = NULL
)
}
\arguments{
\item{refdata}{Standard population dataset you want to use.}

\item{newGroups}{Create a list of new age groups you want to create.}

\item{event}{Column in refdata with outcome counts.}

\item{age}{Column in refdata with age values.}

\item{pop}{Column in refdata with population counts, preferably in person-time.}

\item{strata}{Column or columns to stratify by.}
}
\value{
Table
}
\description{
Create new merged age groups
}
\examples{
\donttest{


esp2013|> dplyr::glimpse()

merged_data <- mergeAgeGroups(esp2013, c("0-19", "20-64", "65-150"))

merged_data |> dplyr::glimpse()

}

}
