% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopTidy.R
\name{omopTidy}
\alias{omopTidy}
\title{Tidy incidence results.}
\usage{
omopTidy(data)
}
\arguments{
\item{data}{A data frame with incidence estimates produced by the
function \code{estimateIncidence()} in the IncidencePrevalence package.}
}
\value{
Data frame
}
\description{
Tidy summarised results produced by \code{IncidencePrevalence}.
}
\examples{
# example code
con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomiaDir("GiBleed"))

cdm <- CDMConnector::cdmFromCon(con, cdmSchema = "main", writeSchema = "main")

cdm$gibleed <- cdm |>
CohortConstructor::conceptCohort(conceptSet = list("gibleed" = 192671L),
name = "gibleed")

cdm <- IncidencePrevalence::generateDenominatorCohortSet(
cdm = cdm,
name = "denominator_esp",
ageGroup = ageGroups(name = "esp2013"),
sex = "Both",
cohortDateRange = as.Date(c("2012-01-01", "2022-01-01"))
)

inc_data_esp <- IncidencePrevalence::estimateIncidence(
cdm = cdm,
denominatorTable = "denominator_esp",
outcomeTable = "gibleed",
interval = "years",
repeatedEvents = TRUE,
outcomeWashout = 30,
completeDatabaseIntervals = FALSE
)

inc_tidy_esp <- omopTidy(inc_data_esp)

}
