% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPopulation.R
\name{standardPopulation}
\alias{standardPopulation}
\title{Standard population}
\usage{
standardPopulation(region = "Europe")
}
\arguments{
\item{region}{Region of standard population. Can be either 'Europe' or 'World'.}
}
\value{
Tibble
}
\description{
Standard population
}
\examples{
\donttest{
standard_data <- standardPopulation(region = "Europe")
}

}
