% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageGroups.R
\name{ageGroups}
\alias{ageGroups}
\title{Age groups}
\usage{
ageGroups(name = "esp2013")
}
\arguments{
\item{name}{Dataset - One of "esp2013" or ""wsp2025".}
}
\value{
List
}
\description{
Age groups
}
\examples{
ageGroups(name = "esp2013")
}
