% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAgeGroups.R
\name{mergeAgeGroups}
\alias{mergeAgeGroups}
\title{Create new merged age groups}
\usage{
mergeAgeGroups(data, newGroups)
}
\arguments{
\item{data}{Standard population dataset you want to use.}

\item{newGroups}{Create a list of new age groups you want to create}
}
\value{
Table
}
\description{
Create new merged age groups
}
\examples{
\donttest{

data <- standardPopulation(name = "esp2013")

data |> dplyr::glimpse()

merged_data <- mergeAgeGroups(data, c("0-19", "20-64", "65-150"))

merged_data |> dplyr::glimpse()

}
}
