% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAgeGroups.R
\name{mergeAgeGroups}
\alias{mergeAgeGroups}
\title{Create new merged age groups}
\usage{
mergeAgeGroups(refdata, newGroups, age = "age_group", pop = "pop")
}
\arguments{
\item{refdata}{Standard population dataset you want to use.}

\item{newGroups}{Create a list of new age groups you want to create.}

\item{age}{Column in refdata with age values.}

\item{pop}{Column in refdata with population counts.}
}
\value{
Table
}
\description{
Create new merged age groups
}
\examples{
\donttest{

refdata <- standardPopulation(name = "esp2013")

refdata |> dplyr::glimpse()

merged_data <- mergeAgeGroups(refdata, c("0-19", "20-64", "65-150"))

merged_data |> dplyr::glimpse()

}

}
