% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopTidy.R
\name{omopTidy}
\alias{omopTidy}
\title{Tidy incidence results.}
\usage{
omopTidy(data)
}
\arguments{
\item{data}{A data frame with incidence estimates produced by the
function \code{estimateIncidence()} in the IncidencePrevalence package.}
}
\value{
Data frame
}
\description{
Tidy summarised results produced by \code{IncidencePrevalence}.
}
