% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPopulation.R
\name{standardPopulation}
\alias{standardPopulation}
\title{Standard population}
\usage{
standardPopulation(name = "esp2013", ageType = "narrow")
}
\arguments{
\item{name}{Dataset to use. One of "esp2013" or "wsp2025".}

\item{ageType}{Choose whether to use broad or narrow age groups.}
}
\value{
Tibble
}
\description{
Standard population
}
