% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPopulation.R
\name{standardPopulation}
\alias{standardPopulation}
\title{Standard population}
\usage{
standardPopulation(name = "esp2013")
}
\arguments{
\item{name}{Dataset to use. One of "esp2013" or "wsp2025".}
}
\value{
Tibble
}
\description{
Standard population
}
\examples{
\donttest{
standard_data <- standardPopulation("esp2013")
}

}
