% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPopulation.R
\name{standardPopulation}
\alias{standardPopulation}
\title{Standard population}
\usage{
standardPopulation(name = "esp2013")
}
\arguments{
\item{name}{Dataset to use. One of "esp2013", "wsp2025", "esp2013_by_sex" or "wsp2025_by_sex".}
}
\value{
Tibble
}
\description{
Standard population
}
\examples{
standardPopulation()
}
